+++ lib/src/common/user/user_manager.dart
@@ -1,15 +1,4 @@
-import 'package:wiredash/src/common/utils/device_info.dart';
-
 class UserManager {
-  UserManager() {
-    DeviceInfo.getDeviceID().then((id) => deviceId = id);
-  }
-
-  String appVersion;
-  String deviceId;
   String userId;
   String userEmail;
-
-  Map<String, dynamic> get deviceInfo =>
-      DeviceInfo.generate(appVersion, deviceId);
 }